changes in header and sidebar
