+++ Frontend/angular-app/src/app/layouts/new-container/new-container.component.scss
@@ -1,139 +1,7 @@
-@import "../../../assets/scss/variables";
-@import "../../../assets/scss/mixins";
-
 .app-container {
-  width: 100%;
-  height: 100%;
   display: flex;
   flex-direction: row;
-  position: relative;
+  height: 100vh;
   overflow: hidden;
-
-  .container-sidebar {
-    grid-area: sidebar;
-    width: $sidebar-width;
-    height: 100%;
-    margin: auto;
-    transition: width 0.15s;
-  }
-
-  .container-sidebar-minimized {
-    grid-area: sidebar;
-    width: $sidebar-width-minimized;
-    height: 100%;
-    margin: 0;
-  }
-
-  .app-container-wrapper {
-    display: grid;
-    height: 100%;
-    width: 100%;
-    min-height: 100vh;
-    max-height: 100vh;
-    overflow: auto;
-    grid-template-areas:
-                         "view";
-    grid-template-columns:  1fr;
-    grid-template-rows: 1fr;
-
-    .help-view-btn {
-      position: absolute;
-      left: 5px;
-      top: 5px;
-    }
-
-    .container-header {
-      grid-area: header;
-      width: 100%;
-      height: 100%;
-    }
-
-
-    .container-view {
-      grid-area: view;
-      width: 100%;
-      height: 100%;
-      max-height: 100%;
-      max-width: 100%;
-      overflow: hidden;
-      margin: auto;
-      display: grid;
-      grid-template-rows: $tabs-height 1fr;
-      padding: 5px 0 0 0;
-
-        .route-content {
-          width: 100%;
-          overflow: hidden;
-          height: 100%;
-          max-width: 100vw;
-          background-color: rgba(color(primary, 50), 0.7);
-          position: relative;
-
-          .tab-spinner {
-            z-index: 9999;
-            position: absolute;
-            left: 0;
-            top: 0;
-            background: rgba(black, .3);
-            display: grid;
-            align-items: center;
-            justify-content: center;
-            width: 100%;
-            height: 100%;
-          }
-
-        //.tab-content {
-        //  margin: auto;
-        //  opacity: 1 !important;
-        //  transition: 0.5s all;
-        //  width: 100%;
-        //  max-width: 100%;
-        //  height: calc(100vh - #{$tabs-height} - #{$header-height}) !important;
-        //  max-height: calc(100vh - #{$tabs-height} - #{$header-height}) !important;
-        //
-        //  ::ng-deep > :first-child {
-        //    height: calc(100vh - #{$tabs-height} - #{$header-height}) !important;
-        //    max-height: calc(100vh - #{$tabs-height} - #{$header-height}) !important;
-        //    overflow-y: auto !important;
-        //    margin: auto !important;
-        //    padding: 1rem;
-        //    display: block;
-        //    align-content: center;
-        //  }
-        //
-        //}
-
-          .tab-content {
-            margin: auto;
-            opacity: 1 !important;
-            transition: 0.5s all;
-            width: 100%;
-            max-width: 100%;
-            height: calc(100vh - #{$tabs-height}) !important;
-            max-height: calc(100vh - #{$tabs-height}) !important;
-
-            ::ng-deep > :first-child {
-              height: calc(100vh - #{$tabs-height}) !important;
-              max-height: calc(100vh - #{$tabs-height}) !important;
-              padding: 1rem;
-              overflow-y: auto !important;
-              margin: auto !important;
-              display: block;
-              /*align-content: center;*/
-              z-index: 1;
-            }
-          }
-
-
-        .hideTabContent {
-          opacity: 0 !important;
-          transition: 0s opacity !important;
-          z-index: -1 !important;
-          position: absolute;
-          right: -200vw;
-          overflow: hidden;
-        }
-      }
-    }
-  }
+  position: relative;
 }
