+++ Frontend/angular-app/src/app/layouts/new-container/sidebar/sidebar.component.html
@@ -1,36 +1,31 @@
-<div class="app-sidebar">
+<div>
+  <!-- Backdrop -->
+  <div
+    *ngIf="isToggledSubmenu"
+    class="sidebar-backdrop"
+    (click)="isToggledSubmenu=false"
+  >
+  </div>
 
-      <!-- eefa logo  -->
-      <div class="logo-container">
-          <svg class="logo" xmlns="http://www.w3.org/2000/svg" width="107.378" height="21.931"
-               viewBox="0 0 107.378 21.931">
-              <g id="Group_544" data-name="Group 544" transform="translate(-1722 -21.035)">
-                  <g id="Group_288" data-name="Group 288" transform="translate(-13)">
-                      <path id="Path_118" data-name="Path 118"
-                            d="M280.051,21.667h-3.219c-.431,0-1.074-.266-.492-1.552,2.125-4.692,8.5-18.492,9.139-19.852.142-.3.064-.266.391-.266h6.652c.119,0,.09-.009.148.1q5.423,10.425,10.849,20.851c.11.208.012.718-.3.718h-8.623a2.044,2.044,0,0,1-.918-.376,30.7,30.7,0,0,1-1.963-3.6c-.113-.223-.055-.2-.3-.2h-8.427c-.353,0-.287-.061-.452.243l-2,3.65c-.185.339-.1.284-.486.284Zm4.414-8.16c-.31.585.009.611.214.611h5.014c.362,0-2.3-6.07-2.307-6.07h-.252C287.062,8.047,284.732,13.006,284.465,13.507Z"
-                            transform="translate(1538.816 21.039)" fill-rule="evenodd" />
-                      <path id="Path_119" data-name="Path 119"
-                            d="M10.191,3.132h11.1c.718-.029.758-3.158-.261-3.132H.01V21.93l22.961-.006c.747,0,.813-3.375,0-3.375l-12.78-.014V11.486h9.26c.984,0,.993-3.419.02-3.416l-9.281.02V3.129Z"
-                            transform="translate(1734.99 21.035)" fill-rule="evenodd" />
-                      <path id="Path_120" data-name="Path 120"
-                            d="M107.051,3.132h11.1c.718-.029.758-3.158-.261-3.132H96.87V21.93l22.961-.006c.747,0,.813-3.375,0-3.375l-12.78-.014V11.486h9.26c.984,0,.993-3.419.02-3.416l-9.281.02V3.129Z"
-                            transform="translate(1666.17 21.035)" fill-rule="evenodd" />
-                      <path id="Path_121" data-name="Path 121"
-                            d="M203.911,3.132h11.1c.718-.029.758-3.158-.261-3.132H193.73V21.93l10.184-.006V11.489h9.26c.984,0,.993-3.419.02-3.416l-9.281.02V3.132Z"
-                            transform="translate(1597.348 21.035)" fill-rule="evenodd" />
-                  </g>
-              </g>
-          </svg>
-      </div>
+  <div class="app-sidebar" [class.app-sidebar-minimized]="isToggled">
 
-  <!--  side items  -->
-  <nav>
-    <app-sidebar-item [isToggled]="isMinimized"
-                      (sidebarItemClick)="itemClicked($event)" [sidebarItem]="baseNavigation"
-                      (collapseSiblings)="collapseOtherSiblings($event)"
-                      *ngFor="let baseNavigation of navigations"
-                      [currentRoute]="currentRoute"
-                     ></app-sidebar-item>
-<!--    <mat-icon class="me-auto toggle-menu-icon item-icon" (click)="toggled.emit(isMinimized)">menu</mat-icon>-->
-  </nav>
+    <!-- eefa logo  -->
+    <div style="min-height: 80px!important;"
+         class="d-flex align-items-center justify-content-center fw-bold fs-4 border-bottom border-secondary text-white">
+      EEFA
+    </div>
+
+    <!--  side items  -->
+    <nav>
+      <app-sidebar-item [isToggled]="isToggled"
+                        [isToggledSubmenu]="isToggledSubmenu"
+                        (sidebarItemClick)="itemClicked($event)" [sidebarItem]="baseNavigation"
+                        (collapseSiblings)="collapseOtherSiblings($event)"
+                        *ngFor="let baseNavigation of navigations"
+                        [currentRoute]="currentRoute"
+                        (unMinimize)="isToggledSubmenu = $event"
+      ></app-sidebar-item>
+      <!--    <mat-icon class="me-auto toggle-menu-icon item-icon" (click)="toggled.emit(isMinimized)">menu</mat-icon>-->
+    </nav>
+  </div>
 </div>
